up: update dependencies and disgo * reorganize GD utils to separated folder * fix and reimplement some things * some additions to command handler and some tests
