+++ commands/commands.go
@@ -1,9 +1,11 @@
 package commands
 
 import (
+	"github.com/dexslender/orb/commands/about"
 	"github.com/dexslender/orb/orb"
 	"github.com/disgoorg/disgo/discord"
 	"github.com/disgoorg/disgo/handler"
+	// "github.com/disgoorg/disgo/handler/middleware"
 )
 
 var List = []discord.ApplicationCommandCreate{
@@ -15,7 +17,7 @@ var List = []discord.ApplicationCommandCreate{
 }
 
 func Setup(o *orb.Orb, router handler.Router) {
-	router.Command("/about/{sub}", runAbout(o))
+	// router.Use(middleware.Logger)
 	router.Command("/gd/{sub}", runGD)
 	router.Command("/setup/{sub}", runSetup)
 
@@ -24,6 +26,10 @@ func Setup(o *orb.Orb, router handler.Router) {
 		r.Command("/ping", runPing)
 	})
 
+	router.Group(func(r handler.Router) {
+		r.Command("/about/{sub}", runAbout(o))
+		r.ButtonComponent("/about/guild/{guildId}", about.UpdateWithGuildStats)
+	})
 
 	router.ButtonComponent("/ping/refresh", runPingRefresh)
 }
