+++ commands/gd/getuser.go
@@ -3,7 +3,7 @@ package gd
 import (
 	"io"
 
-	"github.com/dexslender/orb/util"
+	"github.com/dexslender/orb/util/gd"
 	"github.com/disgoorg/disgo/discord"
 	"github.com/disgoorg/disgo/handler"
 	"github.com/k0kubun/pp/v3"
@@ -23,29 +23,30 @@ var GetUserCommand = discord.ApplicationCommandOptionSubCommand{
 
 func GetUserRun(ctx* handler.CommandEvent) error {
 	query := ctx.SlashCommandInteractionData().String("query")
-	res, err := util.GDClient.Request(
-		util.Users,
-		util.UsersParams{
+	res, err := gd.GDClient.Request(
+		gd.Users,
+		gd.UsersParams{
 			Query: query,
-			Secret: util.COMMON_KEY,
+			Secret: gd.COMMON_KEY,
 		},
 	)
 	if err != nil { return err }
 	defer res.Body.Close()
+
 	data, err := io.ReadAll(res.Body)
 	if err != nil { return err }
-	if user := util.DecodeGDData[util.PartialUser](string(data)); user != nil {
+	if user := gd.DecodeGDData[gd.PartialUser](string(data)); user != nil {
 		fuser := pp.Sprint(user) 
 		return ctx.CreateMessage(discord.NewMessageCreateBuilder().
 			SetIsComponentsV2(true).
 			AddComponents(discord.NewTextDisplayf("```ansi\n%s```", fuser)).
 		Build())
+	} else {
+		return ctx.CreateMessage(discord.NewMessageCreateBuilder().
+			SetIsComponentsV2(true).
+			AddComponents(discord.NewContainer(
+				discord.NewTextDisplay(":no_entry_sign: _Not Found_"),
+			)).
+		Build())
 	}
-	return ctx.CreateMessage(discord.NewMessageCreateBuilder().
-		SetIsComponentsV2(true).
-		AddComponents(discord.NewContainer(
-			discord.NewTextDisplay(":no_entry_sign: _Not Found_"),
-		)).
-	Build())
 }
-// xd
