+++ util/gd/gd_test.go
@@ -1,4 +1,4 @@
-package util
+package gd
 
 import (
 	"io"
