+++ util/gd/structs.go
@@ -1,56 +1,4 @@
-package util
-
-import (
-	"reflect"
-	"strconv"
-	"strings"
-)
-
-func DecodeGDData[T any](data string) *T {
-	spl := strings.Split(data, "#")
-	RobFormat := strings.Split(spl[0], ":")
-	if len(RobFormat) <= 1 { return nil }
-
-	gd := make(map[string]string)
-	for i, key := range RobFormat {
-		if i&1 == 0 {
-			gd[key] = RobFormat[i+1]
-		}
-	}
-
-	var target T
-	tS := reflect.TypeOf(target)
-	if tS.Kind() != reflect.Struct {
-		return nil
-	}
-	vS := reflect.ValueOf(&target)
-	updateStruct(gd, tS, vS.Elem())
-	return &target
-}
-
-func updateStruct(data map[string]string, t reflect.Type, v reflect.Value) {
-	for i := range t.NumField() {
-		if tag, ok := t.Field(i).Tag.Lookup("prop"); ok {
-			field := v.Field(i)
-			if field.CanSet() {
-				robvalue := data[tag]
-				switch field.Kind() {
-				case reflect.String:
-					field.SetString(robvalue)
-				case reflect.Int:
-					vi, _ := strconv.Atoi(robvalue)
-					field.Set(reflect.ValueOf(vi))
-				case reflect.Bool:
-					vb, _ := strconv.ParseBool(robvalue)
-					field.SetBool(vb)
-				case reflect.Struct:
-					tfield := field.Type()
-					updateStruct(data, tfield, field)
-				}
-			}
-		}
-	}
-}
+package gd
 
 type PartialUser struct {
 	UserName      string `prop:"1"`
